requirements.txt
# 🦮 Blind Assistant - Python Dependencies
# Essential packages for AI assistive technology

# Computer Vision and Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
pandas==2.0.3

# Face Recognition
face-recognition==1.3.0
dlib==19.24.2

# Object Detection
ultralytics==8.0.196

# OCR Text Recognition
pytesseract==0.3.10

# Text-to-Speech
pyttsx3==2.90

# Color Analysis
webcolors==1.13

# GUI Framework (usually pre-installed)
# tkinter - comes with Python

# Additional utilities
matplotlib==3.7.2
cmake==3.27.4

# Note: Some packages may require system dependencies:
# - tesseract-ocr (sudo apt-get install tesseract-ocr)
# - cmake (for dlib compilation)
# - python3-tk (for tkinter GUI)
